backend/requirements.txt

Django==3.2
djangorestframework
djoser
Pillow=10.0.0
sorl-thumbnail
webcolors
reportlab
json2pdf
